Add python version NNS Contract
