--- nns/python/compile.py
@@ -1,2 +1,0 @@
-from boa.compiler import Compiler
-Compiler.load_and_save('domain.py')
